pyproject.toml
[project]
name = "mostra-profissoes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "plotly>=6.3.1",
    "streamlit>=1.50.0",
    "streamlit-image-coordinates>=0.4.0",
    "streamlit-plotly-events>=0.0.6",
]
